Add files via upload
In this project, we are going one step ahead with GPS and going to track a vehicle using GPS and GSM. This Vehicle Tracking System can also be used for Accident Detection Alert System, Soldier Tracking System and many more, by just making few changes in hardware and software. Tracking of vehicles is a process in which we track the vehicle location in form of Latitude and Longitude (GPS coordinates).